requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.8.0
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.5.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.2.4
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.2
